+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Upcoming
+### Changes
+* Updated Cheaper chains data pack for latest tconstruct
+
 ## 4.11.15
 ### Additions
 * Combat.Shields feature